Jeg har rykket regenerateSlots-metoden fra ActivityService til SlotServiceImpl og lavet en test til den. Testen er lidt wack men det må vi leve med xD
