+++ src/main/java/org/example/adventurexp/service/SlotServiceImpl.java
@@ -83,6 +83,22 @@ public class SlotServiceImpl implements ISlotService {
         return createdSlots;
     }
 
+    public void regenerateFutureSlots(Long activityId, LocalDate fromDate) {
+        Activity activity = activityRepository.findById(activityId).orElseThrow(() -> new IllegalArgumentException("Activity not found: " + activityId));
+
+        // Delete existing slots from 'fromDate' onwards
+        LocalDateTime fromDateTime = fromDate.atStartOfDay();
+        List<TimeSlot> slotsToDelete = timeSlotRepository.findByActivityAndStartsAtAfter(activity, fromDateTime);
+        timeSlotRepository.deleteAll(slotsToDelete);
+
+        // Regenerate slots (daily from 08 to 22)
+        LocalDate toDate = fromDate.plusDays(30); // Regenerate slots for the next 30 days
+        LocalTime openTime = LocalTime.of(8, 0); // opens at 08:00
+        LocalTime closeTime = LocalTime.of(22, 0); // closes at 22:00
+
+        generateSlots(activityId, fromDate, toDate, openTime, closeTime);
+    }
+
 
 
     // ---- Hjælpemetode ----
